separated server properties into Environment for better configurability
